refactor transfer metadata
