+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/core/DecodeTransferMetadataServletFilter.java
@@ -13,53 +13,50 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
+ *
  */
 
-package com.tencent.cloud.common.metadata.filter.web;
+package com.tencent.cloud.metadata.core;
 
+import java.io.IOException;
 import java.io.UnsupportedEncodingException;
 import java.net.URLDecoder;
 import java.util.Map;
+
+import javax.servlet.FilterChain;
+import javax.servlet.ServletException;
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
 
 import com.tencent.cloud.common.constant.MetadataConstant;
 import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import com.tencent.cloud.common.util.JacksonUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import reactor.core.publisher.Mono;
 
-import org.springframework.core.Ordered;
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.server.reactive.ServerHttpRequest;
+import org.springframework.core.annotation.Order;
 import org.springframework.util.StringUtils;
-import org.springframework.web.server.ServerWebExchange;
-import org.springframework.web.server.WebFilter;
-import org.springframework.web.server.WebFilterChain;
+import org.springframework.web.filter.OncePerRequestFilter;
 
 /**
  * Filter used for storing the metadata from upstream temporarily when web application is
- * REACTIVE.
+ * SERVLET.
  *
  * @author Haotian Zhang
  */
-public class MetadataReactiveFilter implements WebFilter, Ordered {
+@Order(MetadataConstant.OrderConstant.WEB_FILTER_ORDER)
+public class DecodeTransferMetadataServletFilter extends OncePerRequestFilter {
 
 	private static final Logger LOG = LoggerFactory
-			.getLogger(MetadataReactiveFilter.class);
+			.getLogger(DecodeTransferMetadataServletFilter.class);
 
 	@Override
-	public int getOrder() {
-		return MetadataConstant.OrderConstant.WEB_FILTER_ORDER;
-	}
-
-	@Override
-	public Mono<Void> filter(ServerWebExchange serverWebExchange,
-			WebFilterChain webFilterChain) {
-		// Get metadata string from http header.
-		ServerHttpRequest serverHttpRequest = serverWebExchange.getRequest();
-		HttpHeaders httpHeaders = serverHttpRequest.getHeaders();
-		String customMetadataStr = httpHeaders
-				.getFirst(MetadataConstant.HeaderName.CUSTOM_METADATA);
+	protected void doFilterInternal(HttpServletRequest httpServletRequest,
+			HttpServletResponse httpServletResponse, FilterChain filterChain)
+			throws ServletException, IOException {
+		// Get custom metadata string from http header.
+		String customMetadataStr = httpServletRequest
+				.getHeader(MetadataConstant.HeaderName.CUSTOM_METADATA);
 		try {
 			if (StringUtils.hasText(customMetadataStr)) {
 				customMetadataStr = URLDecoder.decode(customMetadataStr, "UTF-8");
@@ -74,16 +71,17 @@ public class MetadataReactiveFilter implements WebFilter, Ordered {
 		Map<String, String> upstreamCustomMetadataMap = JacksonUtils
 				.deserialize2Map(customMetadataStr);
 
-		MetadataContextHolder.init(upstreamCustomMetadataMap, null);
+		try {
+			MetadataContextHolder.init(upstreamCustomMetadataMap, null);
 
-		// Save to ServerWebExchange.
-		serverWebExchange.getAttributes().put(
-				MetadataConstant.HeaderName.METADATA_CONTEXT,
-				MetadataContextHolder.get());
-		return webFilterChain.filter(serverWebExchange)
-				.doOnError(throwable -> LOG.error("handle metadata[{}] error.",
-						MetadataContextHolder.get(), throwable))
-				.doFinally((type) -> MetadataContextHolder.remove());
+			filterChain.doFilter(httpServletRequest, httpServletResponse);
+		}
+		catch (IOException | ServletException | RuntimeException e) {
+			throw e;
+		}
+		finally {
+			MetadataContextHolder.remove();
+		}
 	}
 
 }
