+++ spring-cloud-starter-tencent-metadata-transfer/src/main/java/com/tencent/cloud/metadata/core/EncodeTransferMedataScgFilter.java
@@ -13,72 +13,72 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
+ *
  */
 
-package com.tencent.cloud.metadata.core.filter.gateway;
+package com.tencent.cloud.metadata.core;
 
 import java.io.UnsupportedEncodingException;
 import java.net.URLEncoder;
 import java.util.Map;
 
-import com.netflix.zuul.ZuulFilter;
-import com.netflix.zuul.context.RequestContext;
 import com.tencent.cloud.common.constant.MetadataConstant;
 import com.tencent.cloud.common.metadata.MetadataContext;
 import com.tencent.cloud.common.metadata.MetadataContextHolder;
 import com.tencent.cloud.common.util.JacksonUtils;
+import reactor.core.publisher.Mono;
 
+import org.springframework.cloud.gateway.filter.GatewayFilterChain;
+import org.springframework.cloud.gateway.filter.GlobalFilter;
+import org.springframework.core.Ordered;
+import org.springframework.http.server.reactive.ServerHttpRequest;
 import org.springframework.util.CollectionUtils;
+import org.springframework.web.server.ServerWebExchange;
 
-import static org.springframework.cloud.netflix.zuul.filters.support.FilterConstants.RIBBON_ROUTING_FILTER_ORDER;
-import static org.springframework.cloud.netflix.zuul.filters.support.FilterConstants.ROUTE_TYPE;
+import static org.springframework.cloud.gateway.filter.LoadBalancerClientFilter.LOAD_BALANCER_CLIENT_FILTER_ORDER;
 
 /**
- * Zuul filter used for writing metadata in HTTP request header.
+ * Scg filter used for writing metadata in HTTP request header.
  *
  * @author Haotian Zhang
  */
-public class Metadata2HeaderZuulFilter extends ZuulFilter {
+public class EncodeTransferMedataScgFilter implements GlobalFilter, Ordered {
+
+	private static final int METADATA_SCG_FILTER_ORDER = LOAD_BALANCER_CLIENT_FILTER_ORDER
+			+ 1;
 
 	@Override
-	public String filterType() {
-		return ROUTE_TYPE;
+	public int getOrder() {
+		return METADATA_SCG_FILTER_ORDER;
 	}
 
 	@Override
-	public int filterOrder() {
-		return RIBBON_ROUTING_FILTER_ORDER - 1;
-	}
-
-	@Override
-	public boolean shouldFilter() {
-		return true;
-	}
-
-	@Override
-	public Object run() {
-		// get request context
-		RequestContext requestContext = RequestContext.getCurrentContext();
+	public Mono<Void> filter(ServerWebExchange exchange, GatewayFilterChain chain) {
+		// get request builder
+		ServerHttpRequest.Builder builder = exchange.getRequest().mutate();
 
 		// get metadata of current thread
-		MetadataContext metadataContext = MetadataContextHolder.get();
+		MetadataContext metadataContext = exchange
+				.getAttribute(MetadataConstant.HeaderName.METADATA_CONTEXT);
 
 		// add new metadata and cover old
+		if (metadataContext == null) {
+			metadataContext = MetadataContextHolder.get();
+		}
 		Map<String, String> customMetadata = metadataContext
 				.getAllTransitiveCustomMetadata();
 		if (!CollectionUtils.isEmpty(customMetadata)) {
 			String metadataStr = JacksonUtils.serialize2Json(customMetadata);
 			try {
-				requestContext.addZuulRequestHeader(
-						MetadataConstant.HeaderName.CUSTOM_METADATA,
+				builder.header(MetadataConstant.HeaderName.CUSTOM_METADATA,
 						URLEncoder.encode(metadataStr, "UTF-8"));
 			}
 			catch (UnsupportedEncodingException e) {
-				requestContext.addZuulRequestHeader(
-						MetadataConstant.HeaderName.CUSTOM_METADATA, metadataStr);
+				builder.header(MetadataConstant.HeaderName.CUSTOM_METADATA, metadataStr);
 			}
 		}
-		return null;
+
+		return chain.filter(exchange.mutate().request(builder.build()).build());
 	}
 
 }
