+++ spring-cloud-starter-tencent-metadata-transfer/src/test/java/com/tencent/cloud/metadata/config/MetadataTransferAutoConfigurationTest.java
@@ -13,13 +13,14 @@
  * under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR
  * CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
+ *
  */
 
 package com.tencent.cloud.metadata.config;
 
-import com.tencent.cloud.metadata.core.filter.gateway.Metadata2HeaderZuulFilter;
-import com.tencent.cloud.metadata.core.interceptor.Metadata2HeaderFeignInterceptor;
-import com.tencent.cloud.metadata.core.interceptor.Metadata2HeaderRestTemplateInterceptor;
+import com.tencent.cloud.metadata.core.EncodeTransferMedataFeignInterceptor;
+import com.tencent.cloud.metadata.core.EncodeTransferMedataRestTemplateInterceptor;
+import com.tencent.cloud.metadata.core.EncodeTransferMetadataZuulFilter;
 import org.assertj.core.api.Assertions;
 import org.junit.Test;
 
@@ -48,17 +49,17 @@ public class MetadataTransferAutoConfigurationTest {
 					Assertions.assertThat(context).hasSingleBean(
 							MetadataTransferAutoConfiguration.MetadataTransferFeignInterceptorConfig.class);
 					Assertions.assertThat(context)
-							.hasSingleBean(Metadata2HeaderFeignInterceptor.class);
+							.hasSingleBean(EncodeTransferMedataFeignInterceptor.class);
 					Assertions.assertThat(context).hasSingleBean(
 							MetadataTransferAutoConfiguration.MetadataTransferRestTemplateConfig.class);
 					Assertions.assertThat(context)
-							.hasSingleBean(Metadata2HeaderRestTemplateInterceptor.class);
+							.hasSingleBean(EncodeTransferMedataRestTemplateInterceptor.class);
 					Assertions.assertThat(context).hasSingleBean(
-							MetadataTransferAutoConfiguration.MetadataTransferRestTemplateConfig.Metadata2HeaderRestTemplatePostProcessor.class);
+							MetadataTransferAutoConfiguration.MetadataTransferRestTemplateConfig.EncodeTransferMetadataRestTemplatePostProcessor.class);
 					Assertions.assertThat(context).hasSingleBean(
 							MetadataTransferAutoConfiguration.MetadataTransferZuulFilterConfig.class);
 					Assertions.assertThat(context)
-							.hasSingleBean(Metadata2HeaderZuulFilter.class);
+							.hasSingleBean(EncodeTransferMetadataZuulFilter.class);
 					Assertions.assertThat(context).hasSingleBean(
 							MetadataTransferAutoConfiguration.MetadataTransferScgFilterConfig.class);
 					Assertions.assertThat(context).hasSingleBean(GlobalFilter.class);
