+++ spring-cloud-starter-tencent-polaris-discovery/src/main/java/com/tencent/cloud/polaris/PolarisDiscoveryProperties.java
@@ -56,7 +56,7 @@ public class PolarisDiscoveryProperties {
 	/**
 	 * Service name to registry.
 	 */
-	@Value("${spring.cloud.polaris.discovery.service:${spring.application.name:}}")
+	@Value("${spring.cloud.polaris.discovery.service:${spring.cloud.polaris.service:${spring.application.name:}}}")
 	private String service;
 
 	/**
